+++ docs/Sample-Code/exampleapp/apps/ui/src/_components/UsersMenu/UserIngredients.jsx
@@ -10,8 +10,6 @@ import {
     INGREDIENT_TYPE,
 } from '../../_constants/app_constants.jsx';
 
-// const dbApiService = gs.dbService.dbApiService;
-// const authenticationService = gs.authenticationService.authenticationService;
 const useUser = gs.UserContext.useUser;
 const GenericCrudEditor = gs.genericEditorRfcService.GenericCrudEditor;
 const GetFormData = gs.genericEditorRfcService.GetFormData;
@@ -19,17 +17,15 @@ const GenericSelectGenerator = gs.genericEditorRfcSelector.GenericSelectGenerato
 const GenericSelectDataPopulator = gs.genericEditorRfcSelector.GenericSelectDataPopulator;
 const genericFuncArrayDefaultValue = gs.genericEditorRfcSpecificFunc.genericFuncArrayDefaultValue;
 const console_debug_log = gs.loggingService.console_debug_log;
-// const convertId = gs.idUtilities.convertId;
-// const ACTION_DELETE = gs.generalConstants.ACTION_DELETE;
 const ChatBotButton = gsAi.ChatBotButton;
 
 
 export function UserIngredients_EditorData() {
     // console_debug_log("UserIngredients_EditorData");
     const registry = {
-        "UserIngredients": UserIngredients, 
-        "UserIngredientsDataPopulator": UserIngredientsDataPopulator, 
-        "UserIngredientsValidations": UserIngredientsValidations, 
+        "UserIngredients": UserIngredients,
+        "UserIngredientsDataPopulator": UserIngredientsDataPopulator,
+        "UserIngredientsValidations": UserIngredientsValidations,
         "CALORIE_UNITS": CALORIE_UNITS,
         "SERVING_SIZE_UNITS": SERVING_SIZE_UNITS,
         "INGREDIENT_TYPE": INGREDIENT_TYPE,
@@ -46,7 +42,7 @@ export const UserIngredientsSelect = (props) => {
     // const { currentUserValue } = authenticationService;
     // const user_id_filter = {'user_id': currentUserValue.id}
     const { currentUser } = useUser();
-    const user_id_filter = {'user_id': currentUser.id}
+    const user_id_filter = { 'user_id': currentUser.id }
     console_debug_log("*** UserIngredientsSelect *** | user_id_filter:");
     console_debug_log(user_id_filter);
     return (
@@ -60,10 +56,8 @@ export const UserIngredientsSelect = (props) => {
 }
 
 export const UserIngredientsDataPopulator = () => {
-    // const { currentUserValue } = authenticationService;
-    // const user_id_filter = {'user_id': currentUserValue.id}
     const { currentUser } = useUser();
-    const user_id_filter = {'user_id': currentUser.id}
+    const user_id_filter = { 'user_id': currentUser.id }
     console_debug_log("*** UserIngredientsDataPopulator *** | user_id_filter:");
     console_debug_log(user_id_filter);
     return (
@@ -78,8 +72,11 @@ export const UserIngredientsValidations = (data, editor, action) => {
     // user_ingredients / dishes pre-deletion validations
     return new Promise((resolve, reject) => {
         let resp = genericFuncArrayDefaultValue(data);
-        switch(action) {
-            // Removed on 2023-11-13 and 2023-12-31
+        switch (action) {
+            // //
+            // // Example of pre-deletion validations to avoid deleting a user_ingredient
+            // // that is referenced in user_ingredients_in_user
+            // //
             // case ACTION_DELETE:
             //     const db = new dbApiService({ url: `${editor.dbApiUrl}/user_ingredients_in_user` });
             //     const users_ingredient_id = convertId(data['_id'])