+++ docs/Sample-Code/fastapitemplate/scripts/copy_env_files.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+echo ""
+echo "Copying .env files from:"
+pwd
+echo "--------"
+cp .env ./server/.env
+cp .env ./ui/.env
+cp .env ./mcp-server/.env