+++ docs/Sample-Code/fastapitemplate/ui/src/_components/HomePage/HomePage.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import * as gs from "genericsuite";
+
+const useUser = gs.UserContext.useUser;
+const debug = false;
+
+export const HomePage = () => {
+    const { currentUser } = useUser();
+    return (
+        <gs.HomePage>
+            <>
+                <h2>Hi {currentUser.firstName}!</h2>
+            </>
+        </gs.HomePage>
+    );
+}