+++ scripts/mkdocs_install.sh
@@ -6,15 +6,15 @@ set -e
 
 if [ -d "venv" ]; then
     echo "🔍 venv directory already exists, removing it"
-    rm -rf venv
+    rm -rf .venv
 fi
 
 echo "🔍 Creating venv directory"
-python3 -m venv venv
-source venv/bin/activate
+python3 -m venv .venv
+source .venv/bin/activate
 
 echo "🔍 Installing mkdocs and dependencies"
-pip install mkdocs-material mkdocs-print-site-plugin mkdocs-git-committers-plugin
+pip install mkdocs-material mkdocs-print-site-plugin mkdocs-git-committers-plugin mkdocs-exclude-search
 
 # pip install mkdocs-minify-plugin
 # pip install mkdocs-git-revision-date-localized-plugin